requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-appconf==1.0.5
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-pwa==1.0.10
django-recaptcha==2.0.6
django-redis==5.0.0
django-select2==7.7.3
django-storages==1.12.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.3
Markdown==3.3.4
martor==1.6.4
oauthlib==3.1.1
Pillow==8.4.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
urllib3==1.26.7
